third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-4/styles-edit-property-after-invalid-rule-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that proper source lines are reported for the styles after unrecognizer / invalid selector.

Initial value
[expanded] 
element.style { ()

[expanded] 
#container { (<style>…</style>)
    padding: 15px;

[expanded] 
div { (user agent stylesheet)
    display: block;

After changing property
[expanded] 
element.style { ()

[expanded] 
#container { (<style>…</style>)
    padding: 20px;

[expanded] 
div { (user agent stylesheet)
    display: block;
